Module cumulus_primitives_proof_size_hostfunction::storage_proof_size

Interface that provides access to the current storage proof size.

Should return the current storage proof size if ProofSizeExt is registered. Otherwise, needs to return u64::MAX.

Structs

HostFunctions

Provides implementations for the extern host functions.

Functions

storage_proof_size

Returns the current storage proof size.